Allow for wildcard matches, using directories named __
In many routing schemes, parameters can be mixed into the URL
as well as the query string. I have a usecase where I want to return
the same object every time you hit an endpoint '/users/:id', so this
is my approach.

We just run everything normally as before; if no match is found, we
search the path for any '__' directories and run it again.